server/requirements.txt
openai==1.5.0
langchain==0.0.351
unstructured==0.11.5
tiktoken==0.5.2
pinecone-client==2.2.4
pypdf==3.17.3
sentence-transformers
python-dotenv
flask 
flask-cors
requests